разбил проект на модули и обновил линтер
